add modules folder with different javascript files
